Refactored metadata tab
