+++ frontend/src/js/tabs/metadata/MetadataEditor.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
+import Box from '@material-ui/core/Box';
 import { MetadataDisplay } from 'solarspell-react-lib';
 import MetadataKebabMenu, { MetadataKebabMenuActionProps } from './MetadataKebabMenu';
 import MetadataActionPanel, { MetadataActionPanelActionProps } from './MetadataActionPanel';
+import MetadataTypeAdder, { MetadataTypeAdderActionProps } from './MetadataTypeAdder';
 
 import { Metadata, MetadataType } from '../../types';
 
@@ -9,6 +11,7 @@ type MetadataEditorActionProps = {
     actions: {
         KebabMenu: MetadataKebabMenuActionProps
         ActionPanel: MetadataActionPanelActionProps
+        AddType: MetadataTypeAdderActionProps
     }
 }
 
@@ -18,9 +21,9 @@ type MetadataEditorProps = {
 } & MetadataEditorActionProps
 
 /**
- * This component handles the editing of metadata types and metadata (excluding adding metadata types).
+ * This component handles the editing of metadata types and metadata.
  * @param props The context and various actions for the menus.
- * @returns An accordion containing all metadata types and metadata with additional menus.
+ * @returns A box containing a button for adding and accordions for all metadata types and metadata with additional menus.
  */
 function MetadataEditor({
     metadata,
@@ -28,20 +31,25 @@ function MetadataEditor({
     actions,
 }: MetadataEditorProps): React.ReactElement {
     return (
-        <MetadataDisplay
-            metadata={metadata}
-            metadataTypes={metadataTypes}
-            tableProps={{
-                components: {
-                    KebabMenu: MetadataKebabMenu,
-                    ActionPanel: MetadataActionPanel,
-                },
-                componentProps: {
-                    KebabMenu: actions.KebabMenu,
-                    ActionPanel: actions.ActionPanel,
-                },
-            }}
-        />
+        <Box p={2}>
+            <MetadataTypeAdder
+                onAddType={actions.AddType.onAddType}
+            />
+            <MetadataDisplay
+                metadata={metadata}
+                metadataTypes={metadataTypes}
+                tableProps={{
+                    components: {
+                        KebabMenu: MetadataKebabMenu,
+                        ActionPanel: MetadataActionPanel,
+                    },
+                    componentProps: {
+                        KebabMenu: actions.KebabMenu,
+                        ActionPanel: actions.ActionPanel,
+                    },
+                }}
+            />
+        </Box>
     );
 }
